move example models to tests
